HUM-87: cr fixes
