+++ packages/apps/human-app/frontend/src/modules/worker/components/jobs/available-jobs/components/available-jobs-job-type-filter.tsx
@@ -1,5 +1,6 @@
 /* eslint-disable camelcase --- ... */
 import { useTranslation } from 'react-i18next';
+import { useMemo } from 'react';
 import { Filtering } from '@/shared/components/ui/table/table-header-menu.tsx/filtering';
 import { JOB_TYPES } from '@/shared/consts';
 import { useFilterUpdates } from '@/modules/worker/hooks/use-filter-updates';
@@ -8,10 +9,14 @@ export function AvailableJobsJobTypeFilter({ isMobile = false }) {
   const { t } = useTranslation();
   const { filterParams, updateFilterParams } = useFilterUpdates({ isMobile });
 
-  const filteringOptions = JOB_TYPES.map((jobType) => ({
-    name: t(`jobTypeLabels.${jobType}`),
-    option: jobType,
-  }));
+  const filteringOptions = useMemo(
+    () =>
+      JOB_TYPES.map((jobType) => ({
+        name: t(`jobTypeLabels.${jobType}`),
+        option: jobType,
+      })),
+    [t]
+  );
 
   const handleClear = () => {
     updateFilterParams({ job_type: undefined });
