+++ packages/apps/human-app/frontend/src/modules/worker/components/jobs/available-jobs/components/mobile/available-jobs-table-mobile.tsx
@@ -1,5 +1,4 @@
 import { useTranslation } from 'react-i18next';
-import { type Dispatch, type SetStateAction } from 'react';
 import { Button } from '@/shared/components/ui/button';
 import { FiltersButtonIcon } from '@/shared/components/ui/icons';
 import { useJobsFilterStore } from '@/modules/worker/hooks/use-jobs-filter-store';
@@ -7,11 +6,11 @@ import { EscrowAddressSearchForm } from '@/modules/worker/components/jobs/escrow
 import { AvailableJobsTableJobsListMobile } from '@/modules/worker/components/jobs/available-jobs/components/mobile/available-jobs-table-jobs-list-mobile';
 
 interface AvailableJobsTableMobileProps {
-  setIsMobileFilterDrawerOpen: Dispatch<SetStateAction<boolean>>;
+  handleOpenMobileFilterDrawer: () => void;
 }
 
 export function AvailableJobsTableMobile({
-  setIsMobileFilterDrawerOpen,
+  handleOpenMobileFilterDrawer,
 }: AvailableJobsTableMobileProps) {
   const { t } = useTranslation();
   const { setSearchEscrowAddress } = useJobsFilterStore();
@@ -27,9 +26,7 @@ export function AvailableJobsTableMobile({
       />
       <Button
         fullWidth
-        onClick={() => {
-          setIsMobileFilterDrawerOpen(true);
-        }}
+        onClick={handleOpenMobileFilterDrawer}
         sx={{
           marginBottom: '32px',
           marginTop: '21px',
